refactor(TransportationNetworkCompanyDataSource): Move up getToken method, create abstract auth conn
