+++ src/main/java/org/opentripplanner/updater/transportation_network_company/OAuthToken.java
@@ -1,5 +1,6 @@
 package org.opentripplanner.updater.transportation_network_company;
 
+import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.io.IOException;
@@ -26,6 +27,7 @@ public class OAuthToken {
     public OAuthToken(HttpURLConnection connection) throws IOException {
         // send request and parse response
         ObjectMapper mapper = new ObjectMapper();
+        mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
         InputStream responseStream = connection.getInputStream();
         OAuthAuthenticationResponse response = mapper.readValue(responseStream, OAuthAuthenticationResponse.class);
         responseStream.close();
@@ -35,7 +37,6 @@ public class OAuthToken {
         tokenExpirationTime.setTime(tokenExpirationTime.getTime() + (response.expires_in - 60) * 1000L);
     }
 
-
     /**
      * Checks if a new token needs to be obtained.
      */
@@ -44,7 +45,7 @@ public class OAuthToken {
     }
 
     /**
-     * Used for testing purposes only
+     * Used for testing purposes only.
      */
     public void makeTokenExpire() {
         tokenExpirationTime = Date.from(Instant.now().minusSeconds(1));
